+++ programs/subprojects/mouse_driver/mouse_driver_messages.h
@@ -7,8 +7,8 @@
 typedef struct mouse_packet_msg {
     uint32_t event;
     int8_t status;
-    int8_t rel_x;
-    int8_t rel_y;
+    int16_t rel_x;
+    int16_t rel_y;
     int8_t rel_z;
 } mouse_packet_msg_t;
